Bug 879011 - Turn some ints into floats to avoid rounding error. r=kentuckyfriendtakahe
UltraBlame original commit: 78ce97fe40bc4a52405fe7e32541e8647a3e3e30

+++ gfx/layers/ipc/Axis.h
@@ -143,7 +143,7 @@ public:
 
 
 
-  Overscroll ScaleWillOverscroll(float aScale, int32_t aFocus);
+  Overscroll ScaleWillOverscroll(float aScale, float aFocus);
 
   
 
@@ -153,7 +153,7 @@ public:
 
 
 
-  float ScaleWillOverscrollAmount(float aScale, int32_t aFocus);
+  float ScaleWillOverscrollAmount(float aScale, float aFocus);
